uint32_t

+++ firmware/stepper_nano_zero/stepper_controller.h
@@ -29,11 +29,11 @@ class filter
 {
   public:
     filter();
-    filter(int16_t x);//non 0 initialisation
+    filter(int32_t x);//non 0 initialisation
   private:
-    int16_t v[3];
+    int32_t v[3];
   public:
-    short step(int16_t x);
+    short step(int32_t x);
 };
 
 typedef enum {
